@@ -13,16 +13,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
+    "@angular/common": "^8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/fire": "^5.2.3",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/pwa": "^0.12.4",
-    "@angular/router": "^7.2.2",
-    "@angular/service-worker": "^7.2.2",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/pwa": "~0.803.29",
+    "@angular/router": "^8.2.14",
+    "@angular/service-worker": "^8.2.14",
     "@ionic-native/clipboard": "^5.21.5",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/file": "^5.27.0",
@@ -34,8 +33,8 @@
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
     "@ionic-native/zip": "^5.27.0",
-    "@ionic/angular": "^4.1.0",
-    "@ionic/storage": "^2.2.0",
+    "@ionic/angular": "^5.9.4",
+    "@ionic/storage": "^2.3.1",
     "bootstrap": "^4.4.1",
     "buffer": "^5.2.1",
     "cordova-clipboard": "^1.3.0",
@@ -51,24 +50,25 @@
     "core-js": "^2.5.4",
     "firebase": "^7.6.1",
     "font-awesome": "^4.7.0",
+    "rxjs": "^6.6.7",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/architect": "~0.13.8",
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
-    "@ionic/angular-toolkit": "~1.5.1",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular-devkit/core": "~9.1.15",
+    "@angular-devkit/schematics": "~9.1.15",
+    "@angular/cli": "~8.3.29",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@ionic/angular-toolkit": "~3.0.0",
     "@ionic/lab": "2.0.6",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.12.54",
-    "codelyzer": "~4.5.0",
+    "@types/node": "^13.13.52",
+    "codelyzer": "^5.0.1",
     "cordova-android": "^9.0.0",
     "cordova-plugin-device": "^2.0.2",
     "cordova-plugin-ionic-keyboard": "^2.1.3",
@@ -86,7 +86,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~8.3.0",
     "tslint": "~5.17.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.1"
   },
   "description": "An Ionic project",
   "cordova": {
