@@ -4,10 +4,10 @@ import { PreloadAllModules, RouterModule, Routes, ExtraOptions } from '@angular/
 
 
 const routes: Routes = [
-  { path: '', loadChildren: './tabs/tabs.module#TabsPageModule' },
+  { path: '', loadChildren: () => import('./tabs/tabs.module').then(m => m.TabsPageModule) },
   // { path: 'plan-lectura', loadChildren: './plan-lectura/plan-lectura.module#PlanLecturaPageModule' },
-  { path: 'plan-detalle/:id', loadChildren: './plan-detalle/plan-detalle.module#PlanDetallePageModule' },
-  { path: 'leer-plan/:libro/:capitulo/:versiculo/:versiculoFinal', loadChildren: './leer-plan/leer-plan.module#LeerPlanPageModule' }
+  { path: 'plan-detalle/:id', loadChildren: () => import('./plan-detalle/plan-detalle.module').then(m => m.PlanDetallePageModule) },
+  { path: 'leer-plan/:libro/:capitulo/:versiculo/:versiculoFinal', loadChildren: () => import('./leer-plan/leer-plan.module').then(m => m.LeerPlanPageModule) }
 ];
 const routerOptions: ExtraOptions = {
   scrollPositionRestoration:'enabled',
